Move AoE interfaces to TypeScript definitions file
