+++ src/parser/core/modules/AoE.d.ts
@@ -1,0 +1,17 @@
+import {AbilityEvent} from 'fflogs'
+import Module from 'parser/core/Module'
+
+interface Hit {
+	id: number
+	instance: number
+	times: number
+	amount: number
+}
+
+export interface AoeEvent extends AbilityEvent {
+	hits: Hit[]
+	sourceID: number
+	successfulHit: boolean
+}
+
+export default class AoE extends Module {}
